Criação das rotas iniciais
