+++ package.json
@@ -9,6 +9,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-just-parallax": "^3.1.14",
+    "react-router-dom": "^6.6.0",
     "react-scripts": "5.0.1",
     "react-spring": "^9.6.1",
     "react-star-wars-crawl": "^2.0.0",
