+++ src/pages/Teste.js
@@ -1,0 +1,9 @@
+function Teste(){
+    return (
+        <div>
+            <h1>Funcionou - Home</h1>
+        </div>
+    )
+}
+
+export default Teste;
